Make servername always ci.alphagov
Make the Nginx primary `server_name` always be `ci.alphagov.co.uk` in all
environments (Vagrant and real). Additional hostnames are then appended as
server aliases - either for testing or transition.

In combination with the changes to puppet-nginx the redirect from
HTTP->HTTPS will preserve the hostname if the client has come in on a known
alias.