Namespace all hiera keys by calling class
Use class params and the data binding support to neatly namespace all keys
in hiera. This makes it much clearer where they are being used and maintains
some structure in the YAML files. Uses stdlib to validate data types.

None of these keys appear in ci-deployment so this change will be atomic
within this one commit.