Correct jenkins_{host,server}name typo
The new hiera key added to common in e9f38b5 didn't match the existing param
name. The hiera name actually makes more sense now in relation to the new
`jenkins_serveraliases` param, so align the class to match hiera.

I only noticed this when making the param mandatory in cbd3990. It was
failing back to the default of ''.