[#53144429] Preseed auth for slaves->master
Use a new defined type from alphagov/puppet-jenkins#1 to create a local
Jenkins user account with a preset API token. This is decoupled from
LDAP/oAuth (as all API tokens are), so it will work out-of-the-box.

Created some new credentials for use in Vagrant. It doesn't matter that
these are public. Production credentials will be generated and committed to
gds/ci-deployment, separately.