[#53144429] Don't preset hashed API token
A user's API token is hashed against a seed that is particular to a given
installation of Jenkins. So we can't reliably preset it. Full details in:

alphagov/puppet-jenkins@02e1edcb8dd4a3ac822f7f12b3bbe250cdb931a7

You will need to update hierdata (but not commit it) with the correct
`ui_pass` each time you create a `ci-master-1` node in Vagrant. For real
environments this should stay relatively static.