[#53062579] porting plugins used in jobs
I took the plugins on existing ci and looked to see which
were used by the jobs on the box, for example:
sudo find . -maxdepth 2 -name config.xml -exec grep versionnumber  {} \+

i got the versions of these used by Jenkins on the ci box and put it all
in puppet. Phew.