Speedup innodb at the expense of durability
This disables flushing to disk at the end of each transaction. Instead
the log is written 1/sec. This means that in a db crash up to 1 second
of data could be lost. However on the CI machine the data is not
precious so it doesn't matter if it is lost.