[#59323898] Daily rotate and dateext for CDN logs
Rotate the CDN logs daily, rather than monthly, because they are expected to
grow fast and it matches the rate that the Transition team are expecting to
consume them. The retention period of one year remains the same.

Also append a date to each log when it's rotated. Rather than a rotating
integer, so that it's easy to see what period each log relates to.