Revert "Set up rbenv in logs processing script"
This reverts commit 40a4dfc6b7b2f35bc6c1866635b7ff8c257c2585.

We did this in an attempt to make the logs_processor able to run
this script from a cron job, which otherwise would not set up the
environment correctly for rbenv. However, the script exits immediately
after sourcing rbenv.sh, probably because `rbenv rehash` exits with 1
with the error 'rbenv: cannot rehash: /usr/lib/rbenv/shims/.rbenv-shim exists'
even though it doesn't. Since we don't need this in here right now,
it's better to remove it. Becoming the logs_processor user with
'sudo su - logs_processor' sets up the environment including rbenv
so this script can be run successfully.