Fix tests.
Travis ignores the Gemfile.lock, and we weren't pinning the version of
parallel and parallel_tests.  The api of these gems has changes with
recent versions, which has caused the tests to fail on Travis.

This PR pins the versions of parallel and parallel_tests, and also
updates the test runner to use the new API.