Use Aptly repo for RabbitMQ
Use the GOV.UK Aptly repository for RabbitMQ packages so that we can use
version 3.x. Pin the version to 3.3.4-1 to keep CI builds as consistent
as possible.

Note that `include_src` is set to false because we don't supply source
packages in our Aptly repo.

For more details please see:
https://github.gds/pages/gds/opsmanual/infrastructure/debian-packaging.html

Based on:
https://github.gds/gds/puppet/commit/bece0d52e59b4a1dbe5f587d857e00b3bb4ede5f