Resolve conflict between rabbitmq modules
We had 2 rabbitmq modules in play at once which was causing confusion.
This renames our one to gds_rabbitmq, and makes it a simple wrapper
to make the puppetlabs module install from our apt mirror.

This moves the place this is hooked into the servers to
jenkins_job_support.pp for consistency with other data stores eg mysql.