Fix regex escapes in Puppet RabbitMQ module
Update the `Puppetfile` to use our own fork of the `puppetlabs-rabbitmq`
module which includes a fix to strip extraneous escaping backslashes
from the output of `rabbitmqctl(1)`[1].

The output of `rabbitmqctl(1)` escapes backslashes. This was documented
in the `rabbitmqctl` man page[2] but was removed when the manpage was
converted to DocBook XML[3].

Tested using our `noop_deploy` Fabric task, before this change:

    [ci-slave-4] out: Notice:
    /Stage[main]/Ci_environment::Jenkins_job_support::Rabbitmq/Rabbitmq_user_permissions[content_store@/]/configure_permission:
    current_value ^amq\\.gen.*$, should be ^amq\.gen.*$ (noop)
    [ci-slave-4] out: Notice:
    /Stage[main]/Ci_environment::Jenkins_job_support::Rabbitmq/Rabbitmq_user_permissions[content_store@/]/read_permission:
    current_value ^(amq\\.gen.*|published_documents_test)$, should be
    ^(amq\.gen.*|published_documents_test)$ (noop)
    [ci-slave-4] out: Notice:
    /Stage[main]/Ci_environment::Jenkins_job_support::Rabbitmq/Rabbitmq_user_permissions[content_store@/]/write_permission:
    current_value ^(amq\\.gen.*|published_documents_test)$, should be
    ^(amq\.gen.*|published_documents_test)$ (noop)

With this change, the RabbitMQ permissions are left as-is (a noop).

[1]:
https://github.com/alphagov/puppetlabs-rabbitmq/commit/72785de752337ef050f0ec5462f96d21ac8e8ce6#diff-d41d8cd98f00b204e9800998ecf8427e
[2]:
https://github.com/rabbitmq/rabbitmq-server/blob/16418a9488e15c4d8ef3bfa9fce69190fb8ec796/docs/rabbitmqctl.1.pod#output-escaping
[3]:
rabbitmq/rabbitmq-server@60c14ba#diff-b96d4e7ccd3d984656b8ab3534e6b460