Define all CI machines in Vagrantfile
Add definitions for all of our CI machines to the Vagrantfile, so that
we can test all of these nodes.

Note that I've changed the IP that `ci-management-1` uses so that the
IPs used by the CI slaves are consecutive. This shouldn't have any
effect as the IPs here don't correspond to those used in the real
environment anyway.