Remove hardcoded user
Under Vagrant, we can log in as the `vagrant` user. In a production
environment, individual user accounts will be created by Puppet.

See also this commit, which enables staff members to login using SSH key
authentication as the `ubuntu` user so that Puppet can be run for the
first time (at which point individual user accounts are created and the
`ubuntu` user is disabled):
https://github.gds/gds/ci-deployment/commit/a958d2edb95e5b8cb8f6a4f078d95ec8ab8e7d9d