Upgrade Puppet in Gemfile to version 3.7
Upgrade Puppet and Facter to match the versions supplied by the Vagrant
box templates we're using.

I'm specifying the exact versions so that we can use the exact same
versions as used by the Vagrant boxes.

I have tested this under Vagrant on the `ci-master`, `ci-slave-1`,
`ci-slave-2`, `ci-slave-4` and `transition-logs-1` machine types.