Switch to elasticsearch/elasticsearch puppet module.
This is currently using the gds_operations/elasticsearch module which
has been deprecated.  This switches to the elasticsearch provided
module, which bring this in line with the govuk puppet, and will allow
us to install different ES versions on some of the slaves.

This introduces a gds_elasticsearch module to wrap the upstream module.
This is a cut-down version of the govuk_elasticsearch module from
govuk's puppet.