CI has 8 CPUs, which means we need 8 DBs for EFG
EFG uses https://github.com/grosser/parallel_tests to run its specs,
which defaults to using as many processes as CPUs available.

For each spec process a new DB needs to be created so that the tests
don't step on each other, hence the need for the DBs to match the CPU count.