Extend pact-broker version scheme.
By default pact-broker only supports semver-style versions. We want to
be able to publish pacts for branches as well as specific versions, so
we need to extend this.

Pact-broker has support for extending the version scheme used by
providing a custom version parser[1]. This therefore adds one that does
this to add support for branch names in addition to the numeric
versions.

[1] https://github.com/bethesque/pact_broker/wiki/Configuration#version-parser