Simplify pact-broker upstart config.
Upstart versions after 1.4 (precise ships with 1.5) have `setuid` and
`setgid` directives, so there's no need to use sudo. This will also make
it easier to pass environment variables to the app in future.

Because the output redirection is no longer done by a root process, it's
now necessary for the log files to be owned by the app user.