Add nmap to Jenkins slaves
I plan to use nmap to verify our firewall configuration from an external
network. Add it to the Jenkins slaves so that it can be run from a
Jenkins job.

Add a sudoers config file so that jenkins can run nmap as root,
which is required for the types of scans we're doing.