Create new slave with newer version of PostgreSQL
Add the apt source to enable install of newer version of PostgreSQL.

The module does not handle upgrading PostgreSQL very well on a machine
with an already existing version of the database server. This adds a
brand new slave which has the newer version of PostgreSQL installed.

Also have the newer version of MongoDB.

This also heads towards a naming scheme which doesn't create machines
with sequential numbering so we can get used to an ephermeral type of
environment.